new: ex3.4 authentification

+++ exercises/3.3-4-4b/src/components/App/index.tsx
@@ -1,10 +1,11 @@
 import { useEffect, useState } from "react";
 import { Outlet } from "react-router-dom";
 import NavBar from "./Navbar";
-import type { Movie, MovieContext, NewMovie } from "../../types";
+import type { AuthenticatedUser, MaybeAuthenticatedUser, Movie, MovieContext, NewMovie, User } from "../../types";
 import Header from "../Header";
 import PageTitle from "../PageTitle";
 import Footer from "../Footer";
+import { clearAuthenticatedUser, getAuthenticatedUser, storeAuthenticatedUser } from "../../utils/session";
 
 type Theme = "light" | "dark";
 
@@ -14,6 +15,7 @@ const App = () => {
   const image =
   "https://unsplash.com/photos/a-lamb-logo-on-a-black-background-ze5wHM9kplc";
   
+  const [authenticatedUser, setAuthenticatedUser] = useState<MaybeAuthenticatedUser>(undefined);
   const [theme, setTheme] = useState<Theme>(() => {
     const savedTheme = localStorage.getItem("theme");
 
@@ -35,6 +37,8 @@ const App = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
   useEffect(() => {
     fetchMovies();
+    const authenticatedUser = getAuthenticatedUser();
+    if(authenticatedUser) setAuthenticatedUser(authenticatedUser);
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
@@ -63,11 +67,13 @@ const App = () => {
   };
   const addMovie = async (newMovie: NewMovie) => {
     try{
+      if(!authenticatedUser) throw new Error("You must be authenticated to add a movie");
       const options = {
         method: "POST",
         body: JSON.stringify(newMovie),
         headers: {
           "Content-type":"application/json",
+          Authorization: authenticatedUser.token,
         },
       };
 
@@ -86,8 +92,11 @@ const App = () => {
 
   const deleteMovie = async (id: number) => {
     try{
+      if(!authenticatedUser) throw new Error("you must be authenticated to remove a movie");
       const options = {
         method: "DELETE",
+        headers: {
+          Authorization: authenticatedUser?.token}
       };
       const response = await fetch(`/api/movies/${id}`, options);
 
@@ -101,11 +110,68 @@ const App = () => {
     }
   }
 
+
+  const registerUser = async (newUser:User) =>{
+    try{
+      const options = {
+        method: "POST",
+        body: JSON.stringify(newUser),
+        headers: {
+          "Content-type": 'application/json'
+        },
+      };
+
+      const response = await fetch("/api/auths/register", options);
+      if(!response.ok) throw new Error(`fetch error: ${response.status} : ${response.statusText}`);
+
+      const createdUser: AuthenticatedUser = await response.json();
+
+      setAuthenticatedUser(createdUser);
+      storeAuthenticatedUser(createdUser);
+      console.log("createdUser: ", createdUser);
+    }catch(err){
+      console.error("registerUser::error: ", err);
+      throw err;
+    }
+  };
+
+  const loginUser = async (user: User) => {
+    try{
+      const options = {
+        method: "POST",
+        body: JSON.stringify(user),
+        headers: {
+          "Content-type": "application/json",
+        },
+      };
+
+      const response = await fetch("/api/auths/login", options);
+      if(!response.ok) throw new Error(`fetch error: ${response.status} : ${response.statusText}`);
+
+      const authenticatedUser: AuthenticatedUser = await response.json();
+      console.log("authenticatedUser : ", authenticatedUser);
+
+      setAuthenticatedUser(authenticatedUser);
+      storeAuthenticatedUser(authenticatedUser);
+    }catch(err){
+      console.error("loginUser::error: ", err);
+      throw err;
+    }
+  }
+
+  const clearUser = () => {
+    clearAuthenticatedUser();
+    setAuthenticatedUser(undefined);
+  }
+
   const movieContext: MovieContext = {
     addMovie,
     deleteMovie,
     movies,
     setMovies,
+    registerUser,
+    loginUser,
+    authenticatedUser,
   };
 
   return (
@@ -114,7 +180,7 @@ const App = () => {
         <PageTitle title={pageTitle} />
       </Header>
       <div>
-        <NavBar />
+        <NavBar authenticatedUser={authenticatedUser} clearUser={clearUser}/>
         <Outlet context={movieContext} />
       </div>
       <Footer image={image} theme={theme}>
